requirements.txt
ollama
litellm
langgraph
crewai
dspy-ai
chromadb
llama-index
open-interpreter
pytest
python-dotenv
